requirements/dev.txt
-r base.txt

black==25.1.0
isort==6.0.1
pyinstaller==6.14.2
pytest==8.4.1
pytest-httpx==0.35.0
